Add support for environment variables
Previously all values had to be exact as there was no
support for any inderection. Now, values can be set
to be read from an environment variable by preceding it
with a $.

For example,
term=$TERM
will set the key "term" to the value "xterm" if you are
using a real operating system.